# Speed up serialization of floats in php5 and php7