--- a/src/php7/igbinary.c
+++ b/src/php7/igbinary.c
@@ -804,7 +804,7 @@ inline static int igbinary_serialize_resize(struct igbinary_serialize_data *igsd
 /* {{{ igbinary_serialize8 */
 /** Serialize 8bit value. */
 inline static int igbinary_serialize8(struct igbinary_serialize_data *igsd, uint8_t i) {
-	if (igbinary_serialize_resize(igsd, 1)) {
+	if (UNEXPECTED(igbinary_serialize_resize(igsd, 1))) {
 		return 1;
 	}
 
@@ -815,12 +815,14 @@ inline static int igbinary_serialize8(struct igbinary_serialize_data *igsd, uint
 /* {{{ igbinary_serialize16 */
 /** Serialize 16bit value. */
 inline static int igbinary_serialize16(struct igbinary_serialize_data *igsd, uint16_t i) {
-	if (igbinary_serialize_resize(igsd, 2)) {
+	uint8_t* append_buffer;
+	if (UNEXPECTED(igbinary_serialize_resize(igsd, 2))) {
 		return 1;
 	}
-
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i >> 8 & 0xff);
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i & 0xff);
+	append_buffer = &igsd->buffer[igsd->buffer_size];
+	append_buffer[0] = (uint8_t) (i >> 8 & 0xff);
+	append_buffer[1] = (uint8_t) (i & 0xff);
+	igsd->buffer_size += 2;
 
 	return 0;
 }
@@ -828,14 +830,17 @@ inline static int igbinary_serialize16(struct igbinary_serialize_data *igsd, uin
 /* {{{ igbinary_serialize32 */
 /** Serialize 32bit value. */
 inline static int igbinary_serialize32(struct igbinary_serialize_data *igsd, uint32_t i) {
-	if (igbinary_serialize_resize(igsd, 4)) {
+	uint8_t* append_buffer;
+	if (UNEXPECTED(igbinary_serialize_resize(igsd, 4))) {
 		return 1;
 	}
 
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i >> 24 & 0xff);
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i >> 16 & 0xff);
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i >> 8 & 0xff);
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i & 0xff);
+	append_buffer = &igsd->buffer[igsd->buffer_size];
+	append_buffer[0] = (uint8_t) (i >> 24 & 0xff);
+	append_buffer[1] = (uint8_t) (i >> 16 & 0xff);
+	append_buffer[2] = (uint8_t) (i >> 8 & 0xff);
+	append_buffer[3] = (uint8_t) (i & 0xff);
+	igsd->buffer_size += 4;
 
 	return 0;
 }
@@ -843,18 +848,21 @@ inline static int igbinary_serialize32(struct igbinary_serialize_data *igsd, uin
 /* {{{ igbinary_serialize64 */
 /** Serialize 64bit value. */
 inline static int igbinary_serialize64(struct igbinary_serialize_data *igsd, uint64_t i) {
-	if (igbinary_serialize_resize(igsd, 8)) {
+	uint8_t* append_buffer;
+	if (UNEXPECTED(igbinary_serialize_resize(igsd, 8))) {
 		return 1;
 	}
 
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i >> 56 & 0xff);
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i >> 48 & 0xff);
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i >> 40 & 0xff);
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i >> 32 & 0xff);
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i >> 24 & 0xff);
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i >> 16 & 0xff);
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i >> 8 & 0xff);
-	igsd->buffer[igsd->buffer_size++] = (uint8_t) (i & 0xff);
+	append_buffer = &igsd->buffer[igsd->buffer_size];
+	append_buffer[0] = (uint8_t) (i >> 56 & 0xff);
+	append_buffer[1] = (uint8_t) (i >> 48 & 0xff);
+	append_buffer[2] = (uint8_t) (i >> 40 & 0xff);
+	append_buffer[3] = (uint8_t) (i >> 32 & 0xff);
+	append_buffer[4] = (uint8_t) (i >> 24 & 0xff);
+	append_buffer[5] = (uint8_t) (i >> 16 & 0xff);
+	append_buffer[6] = (uint8_t) (i >> 8 & 0xff);
+	append_buffer[7] = (uint8_t) (i & 0xff);
+	igsd->buffer_size += 8;
 
 	return 0;
 }
@@ -1822,9 +1830,9 @@ inline static uint8_t igbinary_unserialize8(struct igbinary_unserialize_data *ig
 /* {{{ igbinary_unserialize16 */
 /** Unserialize 16bit value. */
 inline static uint16_t igbinary_unserialize16(struct igbinary_unserialize_data *igsd) {
-	uint32_t ret =
-	       ((uint64_t) ((igsd->buffer_ptr[0])) << 8) |
-	       ((uint64_t) ((igsd->buffer_ptr[1])));
+	uint16_t ret =
+	       ((uint16_t) (igsd->buffer_ptr[0]) << 8) |
+	       ((uint16_t) (igsd->buffer_ptr[1]));
 	igsd->buffer_ptr += 2;
 	return ret;
 }
@@ -1833,10 +1841,10 @@ inline static uint16_t igbinary_unserialize16(struct igbinary_unserialize_data *
 /** Unserialize 32bit value. */
 inline static uint32_t igbinary_unserialize32(struct igbinary_unserialize_data *igsd) {
 	uint32_t ret =
-	       ((uint64_t) ((igsd->buffer_ptr[0])) << 24) |
-	       ((uint64_t) ((igsd->buffer_ptr[1])) << 16) |
-	       ((uint64_t) ((igsd->buffer_ptr[2])) << 8) |
-	       ((uint64_t) ((igsd->buffer_ptr[3])));
+	       ((uint32_t) (igsd->buffer_ptr[0]) << 24) |
+	       ((uint32_t) (igsd->buffer_ptr[1]) << 16) |
+	       ((uint32_t) (igsd->buffer_ptr[2]) << 8)  |
+	       ((uint32_t) (igsd->buffer_ptr[3]));
 	igsd->buffer_ptr += 4;
 	return ret;
 }
@@ -1851,7 +1859,7 @@ inline static uint64_t igbinary_unserialize64(struct igbinary_unserialize_data *
 	       ((uint64_t) ((igsd->buffer_ptr[3])) << 32) |
 	       ((uint64_t) ((igsd->buffer_ptr[4])) << 24) |
 	       ((uint64_t) ((igsd->buffer_ptr[5])) << 16) |
-	       ((uint64_t) ((igsd->buffer_ptr[6])) << 8) |
+	       ((uint64_t) ((igsd->buffer_ptr[6])) << 8)  |
 	       ((uint64_t) ((igsd->buffer_ptr[7])) << 0);
 	igsd->buffer_ptr += 8;
 	return ret;