benchmark/serialize-float-array (unrealistic example) sped up from
0.27s to 0.15s with CFLAGS=-O3 (php master branch)
and from 0.304 s to 0.224 s in php 5.6

GCC is better at optimizing append_buffer[0,1,2,3] = val than it is at
optimizing append_buffer[append_offset++] = val

Add a benchmark for serializing packed arrays of floats